another step in refactoring the osd file code
