+++ src/fileio.c
@@ -176,6 +176,96 @@ mame_file *mame_fopen(const char *gamename, const char *filename, int filetype, 
 }
 
 
+/******************************************************************************
+
+File I/O
+
+******************************************************************************/
+static const char* const paths[] = { "raw", "rom", "image", "diff", "samples", "artwork", "nvram", "hi", "hsdb", "cfg", "inp", "memcard", "history", "cheat", "lang", "ctrlr" };
+
+int osd_get_path_count(int pathtype)
+{
+	return 1;
+}
+
+int osd_get_path_info(int pathtype, int pathindex, const char *filename)
+{
+   char buffer[1024];
+   char currDir[1024];
+
+   switch (pathtype)
+   {
+      case FILETYPE_ROM:
+      case FILETYPE_IMAGE:
+         strcpy(currDir, options.libretro_content_path);
+         break;
+      case FILETYPE_IMAGE_DIFF:
+      case FILETYPE_NVRAM:
+      case FILETYPE_HIGHSCORE:
+      case FILETYPE_CONFIG:
+      case FILETYPE_INPUTLOG:
+      case FILETYPE_MEMCARD:
+      case FILETYPE_SAMPLE:
+         /* user generated content goes in mam2003 save directory subfolders */
+         snprintf(currDir, 1024, "%s%s%s%s%s", options.libretro_save_path, path_default_slash(), APPNAME, path_default_slash(), paths[pathtype]);
+         break;
+      default:
+         /* .dat files and additional core content goes in mame2003 system directory */
+         snprintf(currDir, 1024, "%s%s%s", options.libretro_system_path, path_default_slash(), APPNAME);
+   }
+
+   snprintf(buffer, 1024, "%s%s%s", currDir, path_default_slash(), filename);
+
+#ifdef DEBUG_LOG
+   fprintf(stderr, "osd_get_path_info (buffer = [%s]), (directory: [%s]), (path type dir: [%s]), (path type: [%d]), (filename: [%s]) \n", buffer, currDir, paths[pathtype], pathtype, filename);
+#endif
+
+   if (path_is_directory(buffer))
+      return PATH_IS_DIRECTORY;
+   else if (filestream_exists(buffer))
+      return PATH_IS_FILE;
+
+   return PATH_NOT_FOUND;
+}
+
+FILE* osd_fopen(int pathtype, int pathindex, const char *filename, const char *mode)
+{
+   char buffer[1024];
+   char currDir[1024];
+   FILE* out;
+
+   switch (pathtype)
+   {
+      case FILETYPE_ROM:
+      case FILETYPE_IMAGE:
+         strcpy(currDir, options.libretro_content_path);
+         break;
+      case FILETYPE_IMAGE_DIFF:
+      case FILETYPE_NVRAM:
+      case FILETYPE_HIGHSCORE:
+      case FILETYPE_CONFIG:
+      case FILETYPE_INPUTLOG:
+      case FILETYPE_MEMCARD:
+      case FILETYPE_SAMPLE:
+         /* user generated content goes in mam2003 save directory subfolders */
+         snprintf(currDir, 1024, "%s%s%s%s%s", options.libretro_save_path, path_default_slash(), APPNAME, path_default_slash(), paths[pathtype]);
+         break;
+      default:
+         /* .dat files and additional core content goes in mame2003 system directory */
+         snprintf(currDir, 1024, "%s%s%s", options.libretro_system_path, path_default_slash(), APPNAME);
+   }
+
+   snprintf(buffer, 1024, "%s%s%s", currDir, path_default_slash(), filename);
+
+   path_mkdir(currDir);
+
+   out = fopen(buffer, mode);
+
+   return out;
+}
+
+
+
 /***************************************************************************
 	mame_fopen_rom
 ***************************************************************************/
